+++ src/table/table-toolbar/index.js
@@ -1,11 +1,12 @@
 import React from 'react'
+import PropTypes from 'prop-types'
 
 import Toolbar from '@material-ui/core/Toolbar'
 import { makeStyles } from '@material-ui/core/styles'
 
-import Search from './search'
 import Download from './download'
 import Toggle from './toggle'
+import DefaultFilter from '../filters/default-filter'
 
 
 const useStyles = makeStyles(() => ({
@@ -17,6 +18,7 @@ const useStyles = makeStyles(() => ({
 const TableToolbar = ({
   // Search
   preGlobalFilteredRows,
+  globalFilter,
   setGlobalFilter,
   // Toggle
   toggleHideColumn,
@@ -30,10 +32,13 @@ const TableToolbar = ({
 
   return (
     <Toolbar>
-      <Search
-        preGlobalFilteredRows={preGlobalFilteredRows}
-        setGlobalFilter={setGlobalFilter}
-      />
+      {allColumns.some((c) => !c.disableGlobalFilter) && (
+        <DefaultFilter
+          preFilteredRows={preGlobalFilteredRows}
+          setFilter={setGlobalFilter}
+          globalFilter={globalFilter}
+        />
+      )}
       <div className={classes.grow} />
       {allColumns.some((c) => !c.noToggle) && (
         <Toggle
@@ -53,7 +58,9 @@ const TableToolbar = ({
 }
 
 TableToolbar.propTypes = {
-  ...Search.propTypes,
+  preGlobalFilteredRows: PropTypes.array.isRequired,
+  globalFilter: PropTypes.string.isRequired,
+  setGlobalFilter: PropTypes.func.isRequired,
   ...Download.propTypes,
   ...Toggle.propTypes,
 }